+++ Mojo/connect6.mojo
@@ -4,14 +4,14 @@ import values as v
 
 alias max_stones = 6
 
-alias values = List[Float16](
-    Float16(0),
-    Float16(1),
-    Float16(5),
-    Float16(25),
-    Float16(125),
-    Float16(625),
-    inf[DType.float16](),
+alias values = List[Float32](
+    Float32(0),
+    Float32(1),
+    Float32(5),
+    Float32(25),
+    Float32(125),
+    Float32(625),
+    inf[DType.float32](),
 )
 
 alias value_table = v.value_table[max_stones, values]()
